+++ database/migrations/2025_04_26_143537_add_order_number_to_orders_table.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->string('order_number')->unique()->nullable()->after('payment_method');
+            if (!Schema::hasColumn('orders', 'order_number')) {
+                $table->string('order_number')->nullable()->after('payment_method');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->dropColumn('order_number');
+            if (Schema::hasColumn('orders', 'order_number')) {
+                $table->dropColumn('order_number');
+            }
         });
     }
 };
